Change int to size_t in insertion to list

+++ src/list.c
@@ -92,7 +92,7 @@ INTERNAL list_element_s *otrng_list_add(void *data, list_element_s *head) {
   return head;
 }
 
-INTERNAL list_element_s *otrng_list_insert_at_position_n(void *data, int pos,
+INTERNAL list_element_s *otrng_list_insert_at_position_n(void *data, size_t pos,
                                                          list_element_s *head) {
   list_element_s *n = list_new();
   if (!n) {
